code/requirements.txt
python-dotenv
kafka-python
pymongo
